fixes from run-time testing  Committer: Joel Morgan <joel@shop-desktop.(none)> 	modified:   controls/hyd_control.c 	modified:   controls/include/control.h 	modified:   controls/pwm.c 	modified:   include/control_str.h 	modified:   int_arb/candriver.c 	modified:   int_arb/elections.c 	modified:   int_arb/include/candriver.h 	modified:   int_arb/intarb.c 	modified:   int_arb/master.c 	modified:   int_arb/slave.c 	modified:   int_arb/ui.c 	modified:   main.c 	modified:   sensors/adc/adc_main.c 	modified:   sensors/adc/ep93xx_adc_dbes.h 	modified:   sensors/di.c 	modified:   sensors/sensors.c

+++ sensors/adc/ep93xx_adc_dbes.h
@@ -79,7 +79,7 @@ int read_channel(unsigned long adc_page, unsigned short channel)
 	POKE32(adc_page + ADCSWLOCK_OFFSET, UNLOCK_VAL); //unlock the soft lock
 
 	//write ADCSwitch reg to select channel
-	POKE32(adc_page + ADCSWITCH_OFFSET, channel);
+	POKE32(adc_page + ADCSWITCH_OFFSET, (unsigned long)channel);
 
 	while(is_ADC_busy(adc_page)); //poll ADCResult
